ZhonyaS/requirement.txt
SQLAlchemy==2.0.41
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask_SocketIO==5.5.1
icecream==2.1.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Requests==2.32.3
tabulate==0.9.0
